# [Bugfix][Kernel] Use int64_t for indices in fp8 quant kernels