--- a/tests/kernels/test_fp8_quant.py
+++ b/tests/kernels/test_fp8_quant.py
@@ -61,6 +61,7 @@ def test_dynamic_per_tensor_fp8_quant(num_tokens: int, hidden_size: int,
     assert torch.allclose(ref_out.to(dtype=torch.float32),
                           ops_out.to(dtype=torch.float32))
 
+
 # Regression test for a case with large activations where an int32 index cannot
 # represent the number of elements.
 @torch.inference_mode()
@@ -69,8 +70,8 @@ def test_fp8_quant_large(seed: int) -> None:
     torch.random.manual_seed(seed)
     torch.cuda.manual_seed(seed)
 
-    num_tokens = 1024000 # Mistral-Nemo's max_position_embeddings
-    hidden_size = 1152   # Smallest hidden_size to reproduce the error
+    num_tokens = 1024000  # Mistral-Nemo's max_position_embeddings
+    hidden_size = 1152  # Smallest hidden_size to reproduce the error
     dtype = torch.bfloat16
 
     x = torch.rand(num_tokens, hidden_size, dtype=dtype, device="cuda")