@comaniac 2024-07-22
Out of curiosity, do you observe any performance regression with int64 indices? I previously saw an issue in triton reported by someone that since int64 uses 2 registers.