@tlrmchlsmth 2024-07-22
@comaniac I did some quick e2e decode runs -- looks like we may be very slightly slower using int64. But IMO we should still land the PR. Are there any micro-benchmarks for just these kernels?

```
python3 benchmarks/benchmark_latency.py --model neuralmagic/Meta-Llama-3-8B-Instruct-FP8 \
--batch-size 16 --input-len 1 --output-len 512
```

This PR:
```
Avg latency: 3.4161813993162164 seconds
10% percentile latency: 3.3850670562125744 seconds
25% percentile latency: 3.3872151313116774 seconds
50% percentile latency: 3.3914742909837514 seconds
75% percentile latency: 3.451796793495305 seconds
90% percentile latency: 3.4673959348350762 seconds
99% percentile latency: 3.554096885798499 seconds
```

main:
```
Avg latency: 3.4008890725672245 seconds
10% percentile latency: 3.3801775192841887 seconds
25% percentile latency: 3.380847539170645 seconds
50% percentile latency: 3.3822259346488863 seconds
75% percentile latency: 3.431579362368211 seconds
90% percentile latency: 3.4506909909192474 seconds
99% percentile latency: 3.462752147950232 seconds
```