@comaniac 2024-07-22
Thanks for the benchmarking and it seems ok.
For micro-benchmark I feel you could base on `benchmarks/cutlass_benchmarks/w8a8_benchmarks.py`.