@comaniac 2024-07-22
btw do you think it makes sense to have a unit test (if it doesn't take unaccepted long time) to cover this bug?