@tlrmchlsmth 2024-07-22
I added a unit test but going to try to get the memory footprint down as it takes over 16 GB right now